refactor(store): moves store modules to separate files
